+++ CHANGELOG.md
@@ -4,6 +4,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.2.3] - 2021-08-16
+### Added
+- Gaussian process regression (GPR) to belief library
+- Runtime exceptions for pathological error cases such as failed matrix inversion, cholesky decomposition, dimension mismatch
+### Changed
+- Parameter set functions now throw exception rather than return false on bad inputs
+- Updated python belief doc to show GPR usage
+- Updated README
+
 ## [0.2.2] - 2021-08-09
 ### Added
 - Gaussian mixture models (GMM) and Gaussian mixture regression (GMR) to belief library